refactor: apply naming convention
