+++ e2e/fixtures/external-vc.ts
