+++ scripts/EW/create-root.ts
@@ -48,7 +48,7 @@ const rpcUrl = chainConfigs()[CHAIN_ID].rpcUrl;
       roleType: 'app',
       version: 1,
       enrolmentPreconditions: [],
-      fields: [],
+      issuerFields: [],
       metadata: {},
       issuer: { issuerType: 'DID', did: [ENERGYWEB_OWNER] },
     },
@@ -71,7 +71,7 @@ const rpcUrl = chainConfigs()[CHAIN_ID].rpcUrl;
       roleType: 'app',
       version: 1,
       enrolmentPreconditions: [],
-      fields: [],
+      issuerFields: [],
       metadata: {},
       issuer: { issuerType: 'DID', did: [KYC_OWNER] },
     },
