+++ scripts/EW/transfer-root.ts
