+++ src/errors/change-ownership-not-possible.error.ts
