+++ src/errors/deleting-namespace-not-possible.error.ts
