+++ src/errors/ens-type-not-supported.error.ts
@@ -1,4 +1,4 @@
-import { ERROR_MESSAGES } from './ErrorMessages';
+import { ERROR_MESSAGES } from './error-messages';
 
 export class ENSTypeNotSupportedError extends Error {
   constructor() {
