+++ src/errors/index.ts
@@ -1,7 +1,7 @@
-export { ENSTypeNotSupportedError } from './ENSTypeNotSupportedError';
-export { MethodNotAvailableInNodeEnvError } from './MethodNotAvailableInNodeError';
-export { ChangeOwnershipNotPossibleError } from './ChangeOwnershipNotPossibleError';
-export { DeletingNamespaceNotPossibleError } from './DeletingNamespaceNotPossibleError';
-export { ENSOwnerNotValidAddressError } from './ENSOwnerNotValidAddressError';
-export { ERROR_MESSAGES } from './ErrorMessages';
-export { MalformedDIDError } from './MalformedDIDError';
+export { ENSTypeNotSupportedError } from './ens-type-not-supported.error';
+export { MethodNotAvailableInNodeEnvError } from './method-not-available-in-node.error';
+export { ChangeOwnershipNotPossibleError } from './change-ownership-not-possible.error';
+export { DeletingNamespaceNotPossibleError } from './deleting-namespace-not-possible.error';
+export { ENSOwnerNotValidAddressError } from './ens-owner-not-valid-address.error';
+export { ERROR_MESSAGES } from './error-messages';
+export { MalformedDIDError } from './malformed-did.error';
