+++ src/errors/malformed-did.error.ts
