+++ src/errors/method-not-available-in-node.error.ts
