+++ src/errors/unregistered-resolver.error.ts
