+++ src/index.ts
@@ -16,9 +16,9 @@
 
 export * from './init';
 export * from './modules/assets';
-export * from './modules/cacheClient';
+export * from './modules/cache-client';
 export * from './modules/claims';
-export * from './modules/didRegistry';
+export * from './modules/did-registry';
 export * from './modules/domains';
 export * from './modules/messaging';
 export * from './modules/signer';
