+++ src/modules/cache-client/index.ts
@@ -1,0 +1,3 @@
+export * from './cache-client.interface';
+export * from './cache-client.service';
+export * from './cache-client.types';
