--- src/modules/cacheClient/index.ts
@@ -1,3 +1,0 @@
-export * from './ICacheClient';
-export * from './cacheClient.service';
-export * from './cacheClient.types';
