+++ src/modules/did-registry/did-registry.validation.ts
