+++ src/modules/did-registry/did.types.ts
