+++ src/modules/did-registry/index.ts
@@ -1,0 +1,2 @@
+export * from './did-registry.service';
+export * from './did.types';
