--- src/modules/didRegistry/index.ts
@@ -1,2 +1,0 @@
-export * from './didRegistry.service';
-export * from './did.types';
