+++ src/modules/signer/index.ts
@@ -1,8 +1,8 @@
 export * from './signer.service';
 export * from './signer.types';
-export * from './privateKeySigner';
-export * from './metamaskSigner';
-export * from './walletConnectMetamask';
-export * from './walletConnectKms';
-export * from './gnosisSigner';
-export * from './ekcSigner';
+export * from './private-key.signer';
+export * from './metamask.signer';
+export * from './wallet-connect-metamask';
+export * from './wallet-connect-kms';
+export * from './gnosis.signer';
+export * from './ekc.signer';
