+++ src/utils/detect-environment.ts
